// src/app/page/master/map-golonganpegawai-pangkat/map-golonganpegawai-pangkat.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapGolonganpegawaiPangkat {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}